Added automatic special k point generation
If the space group is set or perceived, the special k points
will be automatically set in the yaehmop band dialog. The user
may rearrange them as they please. The special k points are set
to be the primitive special k points, so they may be different
from conventional special k points. Conditions are taken into
automatically.