Added Yaehmop projected DOS calculations
The PDOS options are very similar to the DOS options
except for a few things: there is an added "projections"
text section and a few checkboxes (one about whether or not to
also plot the total DOS, and one about whether or not to integrate
the PDOS).

In the projected DOS text section, the user can enter titles as
lines that begin with '#'. The line immediately following it is
the line that it is a title of. For example:

 # Title
 projection

projections can be 'atom' or 'orbital', and you must then specify
its number and its weight (1.0 is a standard weight).

For example:

 # s orbital
 orbital 1 1.0

When the calculation is performed, a dialog box will appear that serves
as a color legend - as in it informs the user which projections are which
color. If the data is to be integrated, the integrations show up as thinner
lines of the same color as the PDOS for that projection.

Because it can be difficult to keep track of the numberings for the
orbitals and atoms, some buttons were added to assist. By default, the
projections box displays the "Display Atom Projections" selection. This will
separate every atom by type and group them together accordingly.

"Display Orbital Projections" will group together orbitals of the same
type and from the same atom (e. g. all s orbitals from carbon will be
grouped together).

"Display Detailed Orbital Projections" will group together only orbitals
of the same type on an individual atom.