Fixed annoying spg tolerance bug
When performing certain crystallography actions, a dialog would
appear and say that spg detection failed, and it would ask if
the user would like to change the tolerance. If they said "yes",
it would not allow them to change the tolerance and instead pop
up with theh same error message again.

It now properly allows the user to change the tolerance.